Implement Font, for assembling fonts and representing characters.
Define a font using the example `fonts/ascii.yml.example`. This has
slots for all ASCII characters. Use `.` to represent off, and `*` to
represent on. Instead of the YAML array syntax, use a string of another
key for an alias (if there is no lowercase, for example).

To assemble a font from a definition file:

    require './font'
    font = Font.new('./fonts/ascii.yml.example')

To output a character on the screen:

    puts font.text('A')

For a character in dcled format:

    font.dcled('A')

For a character in dcled format as a hash:

    font.dcled_hash('A')